sicfaae-drh/requirements.txt
Django==6.0
python-decouple
django-simple-history
djangorestframework==3.14.0
openpyxl
reportlab
Pillow
requests
